+++ backend/app/Http/Controllers/BookingController.php
@@ -77,7 +77,7 @@ class BookingController extends Controller
             $sizeOfTime = $service->getTimeService($request->service_id) * 4;
             $status = $this->updateShiftStatusAfterBooking($sts, $request->start_time, $sizeOfTime);
             $shift->updateStatusByStylistID($stylist_id, $request->date, $status);
-            return response()->success($newBooking, 'Bạn đã đặt lịch thành công');
+            return response()->success($sentMessage, 'Bạn đã đặt lịch thành công');
         } catch (Exception $e) {
             return response()->exception($e->getMessage(), $e->getCode());
         }
